some reworking of pointings_in_obs.py and pointings.py
