+++ litebird_sim/mapmaking/common.py
@@ -11,7 +11,7 @@ from litebird_sim.coordinates import CoordinateSystem
 from litebird_sim.observations import Observation
 from litebird_sim.mpi import MPI_COMM_GRID
 
-from litebird_sim.pointings import _get_pointings_array, _get_pol_angle
+from litebird_sim.pointings_in_obs import _get_pointings_array, _get_pol_angle
 
 
 # The threshold on the conditioning number used to determine if a pixel
