+++ test/test_hwp_diff_emiss.py
@@ -1,7 +1,7 @@
 import numpy as np
 import litebird_sim as lbs
 from astropy.time import Time
-from litebird_sim.pointings import _get_hwp_angle
+from litebird_sim.pointings_in_obs import _get_hwp_angle
 
 
 def test_add_2f():
